# app-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return current user profile"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "roleId",
            "required": false,
            "in": "query",
            "description": "Filter by Role ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by Role",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "description": "Filter by Department ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by Active status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query for name or email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return users list"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of users (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new user (Admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user detail"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user detail by ID (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deactivated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete/deactivate user (Admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/reset-password": {
      "post": {
        "operationId": "UsersController_resetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset password email sent"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin force reset password for user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/notification-preferences": {
      "patch": {
        "operationId": "UsersController_updateNotificationPreference",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user notification preferences",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "parameters": [
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all notifications as read for current user",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a specific notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "HelpDeskPro API",
    "description": "Platform Helpdesk Internal Berbasis RBAC dan SLA Management",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar URL"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user"
          },
          "password": {
            "type": "string",
            "description": "Initial password",
            "minLength": 8
          },
          "roleId": {
            "type": "string",
            "description": "UUID of the role"
          },
          "departmentId": {
            "type": "string",
            "description": "UUID of the department"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "roleId",
          "departmentId"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "roleId": {
            "type": "string",
            "description": "UUID of the role"
          },
          "departmentId": {
            "type": "string",
            "description": "UUID of the department"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status"
          }
        }
      },
      "UpdateNotificationPreferenceDto": {
        "type": "object",
        "properties": {
          "emailNotificationEnabled": {
            "type": "boolean",
            "description": "Enable or disable email notifications"
          }
        },
        "required": [
          "emailNotificationEnabled"
        ]
      },
...
    }
  }
}